+++ sandbox_marker_verify_linux_test.go
@@ -36,7 +36,7 @@ func TestContainmentNotDisprovedUnderLandlock(t *testing.T) {
 		os.Exit(0)
 	}
 
-	if fd, err := landlockCreateRuleset(landlockAccessFull); err != nil {
+	if fd, err := landlockCreateRuleset(landlockHandledAccess()); err != nil {
 		t.Skipf("landlock unavailable on this kernel: %v", err)
 	} else {
 		_ = syscall.Close(fd)